+++ packages/salmon-price-functions/src/providers/index.ts
@@ -1,3 +1,1 @@
-export * from './tiingo'
-export * from './iex_cloud'
 export * from './finnhubb'
