+++ packages/salmon-provider-iexcloud/README.md
@@ -1,0 +1,5 @@
+# @defichain/salmon-provider-iexcloud
+
+### Config
+
+Requires a iexcloud API token
