+++ tsconfig.build.json
@@ -11,6 +11,8 @@
     {"path": "./packages/salmon-oracles-functions/tsconfig.build.json"},
     {"path": "./packages/salmon-price-functions/tsconfig.build.json"},
     {"path": "./packages/salmon-provider-tiingo/tsconfig.build.json"},
+    {"path": "./packages/salmon-provider-iexcloud/tsconfig.build.json"},
+    {"path": "./packages/salmon-provider-finnhubb/tsconfig.build.json"},
     {"path": "./packages/salmon-lambda-functions/tsconfig.build.json"}
  ]
 }
